deprecated_text_extraction/requirements.txt
kafka-python==2.0.2
python-dotenv==0.21.1
watchdog==2.2.1
PyPDF2==3.0.1
python-docx==0.8.11
docx2txt==0.8
nltk==3.8.1
langdetect==1.0.9
iso-639==0.4.5
transformers==4.26.1
torch==2.0.0
pandas==1.5.3
datasets==2.10.1
scikit-learn==1.2.2
openai==0.27.2
spacy==3.5.1
ipython==8.11.0
PyMuPDF==1.21.1
Unidecode==1.3.6
pdfquery==0.4.3
grobid-client==0.8.3
openapi-python-client==0.13.2
parsr-client==3.2.3
sentence-transformers==2.2.2
